Add sigma0 in lmer() function and propagate down to the Cpp code.
Had to redefine R classes and external.cpp functions to increase number of expected arguments.

Had to redefine classes and functions inside Cpp to include sigma0 in the response module.

I also added sigma0 in the predictive module. But I am guessing that was not actually needed? To explore later.